refactor: rename EnhancedVersionSelectionPage to VersionSelectionPage
This commit moves the enhanced version selection implementation to replace the original VersionSelectionPage component. It removes unused components like landing-page-toggle, standalone-theme-toggle, color-theme, and version-aware-footer. The commit also fixes a typo in the NavigationEventData interface export and removes the duplicate enhanced-error-boundary file.
