--- components/standalone-theme-toggle.tsx
@@ -1,48 +1,0 @@
-"use client"
-
-import { useState, useEffect } from "react"
-import { Moon, Sun } from "lucide-react"
-import { Button } from "@/components/ui/button"
-
-interface StandaloneThemeToggleProps {
-  className?: string
-}
-
-export function StandaloneThemeToggle({ className }: StandaloneThemeToggleProps) {
-  const [mounted, setMounted] = useState(false)
-  const [theme, setTheme] = useState<"light" | "dark">("light")
-
-  useEffect(() => {
-    setMounted(true)
-    // Get initial theme
-    const isDark = document.documentElement.classList.contains("dark")
-    setTheme(isDark ? "dark" : "light")
-  }, [])
-
-  const toggleTheme = () => {
-    const newTheme = theme === "light" ? "dark" : "light"
-    setTheme(newTheme)
-
-    // Update DOM
-    document.documentElement.classList.remove("light", "dark")
-    document.documentElement.classList.add(newTheme)
-
-    // Save preference
-    localStorage.setItem("theme", newTheme)
-  }
-
-  if (!mounted) return null
-
-  return (
-    <Button
-      variant="ghost"
-      size="icon"
-      onClick={toggleTheme}
-      className={className}
-      aria-label={`Switch to ${theme === "dark" ? "light" : "dark"} mode`}
-    >
-      <Sun className="h-[1.2rem] w-[1.2rem] rotate-0 scale-100 transition-all dark:-rotate-90 dark:scale-0" />
-      <Moon className="absolute h-[1.2rem] w-[1.2rem] rotate-90 scale-0 transition-all dark:rotate-0 dark:scale-100" />
-    </Button>
-  )
-}
