+++ components/ui/enhanced-error-boundary.tsx
@@ -4,6 +4,7 @@ import { Component, type ErrorInfo, type ReactNode } from "react"
 import { AlertCircle, RefreshCw } from "lucide-react"
 import { Button } from "@/components/ui/button"
 import { Card, CardContent, CardFooter, CardHeader, CardTitle } from "@/components/ui/card"
+import { cn } from "@/lib/utils"
 
 interface ErrorBoundaryProps {
   children: ReactNode
@@ -35,7 +36,7 @@ export class EnhancedErrorBoundary extends Component<ErrorBoundaryProps, ErrorBo
   }
 
   componentDidCatch(error: Error, errorInfo: ErrorInfo): void {
-    console.error("Error caught by ErrorBoundary:", error, errorInfo)
+    console.error("Error caught by EnhancedErrorBoundary:", error, errorInfo)
     if (this.props.onError) {
       this.props.onError(error, errorInfo)
     }
@@ -54,11 +55,11 @@ export class EnhancedErrorBoundary extends Component<ErrorBoundaryProps, ErrorBo
   render(): ReactNode {
     if (this.state.hasError) {
       if (this.props.fallback) {
-        return this.props.fallback
+        return <>{this.props.fallback}</>
       }
 
       return (
-        <Card className={`border-destructive/50 ${this.props.className || ""}`}>
+        <Card className={cn("border-destructive/50", this.props.className)}>
           <CardHeader className="bg-destructive/10 border-b border-destructive/20">
             <CardTitle className="text-destructive flex items-center gap-2">
               <AlertCircle className="h-5 w-5" />
@@ -80,6 +81,6 @@ export class EnhancedErrorBoundary extends Component<ErrorBoundaryProps, ErrorBo
       )
     }
 
-    return this.props.children
+    return <>{this.props.children}</>
   }
 }
