--- components/ui/enhanced-error_boundary.tsx
@@ -1,86 +1,0 @@
-"use client"
-
-import { Component, type ErrorInfo, type ReactNode } from "react"
-import { AlertCircle, RefreshCw } from "lucide-react"
-import { Button } from "@/components/ui/button"
-import { Card, CardContent, CardFooter, CardHeader, CardTitle } from "@/components/ui/card"
-import { cn } from "@/lib/utils"
-
-interface ErrorBoundaryProps {
-  children: ReactNode
-  fallback?: ReactNode
-  onReset?: () => void
-  onError?: (error: Error, errorInfo: ErrorInfo) => void
-  className?: string
-}
-
-interface ErrorBoundaryState {
-  hasError: boolean
-  error: Error | null
-}
-
-export class EnhancedErrorBoundary extends Component<ErrorBoundaryProps, ErrorBoundaryState> {
-  constructor(props: ErrorBoundaryProps) {
-    super(props)
-    this.state = {
-      hasError: false,
-      error: null,
-    }
-  }
-
-  static getDerivedStateFromError(error: Error): ErrorBoundaryState {
-    return {
-      hasError: true,
-      error,
-    }
-  }
-
-  componentDidCatch(error: Error, errorInfo: ErrorInfo): void {
-    console.error("Error caught by EnhancedErrorBoundary:", error, errorInfo)
-    if (this.props.onError) {
-      this.props.onError(error, errorInfo)
-    }
-  }
-
-  resetErrorBoundary = (): void => {
-    if (this.props.onReset) {
-      this.props.onReset()
-    }
-    this.setState({
-      hasError: false,
-      error: null,
-    })
-  }
-
-  render(): ReactNode {
-    if (this.state.hasError) {
-      if (this.props.fallback) {
-        return <>{this.props.fallback}</>
-      }
-
-      return (
-        <Card className={cn("border-destructive/50", this.props.className)}>
-          <CardHeader className="bg-destructive/10 border-b border-destructive/20">
-            <CardTitle className="text-destructive flex items-center gap-2">
-              <AlertCircle className="h-5 w-5" />
-              Something went wrong
-            </CardTitle>
-          </CardHeader>
-          <CardContent className="pt-4">
-            <div className="text-sm text-muted-foreground">
-              {this.state.error?.message || "An unexpected error occurred"}
-            </div>
-          </CardContent>
-          <CardFooter>
-            <Button variant="outline" size="sm" onClick={this.resetErrorBoundary} className="gap-1">
-              <RefreshCw className="h-3.5 w-3.5" />
-              Try again
-            </Button>
-          </CardFooter>
-        </Card>
-      )
-    }
-
-    return <>{this.props.children}</>
-  }
-}
